# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user": {
      "post": {
        "description": "注册新用户账号，分配默认 user 角色，返回列表行结构。",
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "创建用户",
        "tags": [
          "用户管理"
        ]
      },
      "get": {
        "description": "支持关键字、状态、角色筛选及排序，返回 items 与 pagination。",
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，需与 pageSize 同时传入",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页数量，需与 page 同时传入，最大 100",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字：邮箱、用户名、昵称、手机号",
            "schema": {
              "example": "zhang",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "账号状态，可传单个或数组",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive",
                  "pending",
                  "suspended"
                ]
              }
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "角色 code，可传单个或数组",
            "schema": {
              "example": "admin",
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "enum": [
                "username",
                "email",
                "jobTitle",
                "createdAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "分页查询用户列表",
        "tags": [
          "用户管理"
        ]
      },
      "delete": {
        "description": "逻辑删除指定用户，禁止删除当前登录用户自身。需要二次确认令牌。",
        "operationId": "UserController_removeMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUsersSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListItemArraySuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "批量软删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "description": "按用户 ID 返回完整档案，含角色、权限、组织与安全信息。",
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户 ID（UUID）",
            "schema": {
              "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "获取用户详情",
        "tags": [
          "用户管理"
        ]
      },
      "patch": {
        "description": "按 ID 更新用户资料；若传入 password 将重新哈希存储。",
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户 ID（UUID）",
            "schema": {
              "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "更新用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/user/restore": {
      "patch": {
        "description": "将已逻辑删除的用户恢复为可用状态。",
        "operationId": "UserController_restoreMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUsersSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "恢复成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListItemArraySuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "批量恢复已软删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/user/status": {
      "patch": {
        "description": "批量设置账号状态；禁用账户请使用 suspended，inactive 仅表示未完成激活。",
        "operationId": "UserController_updateStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsersStatusSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListItemArraySuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "批量更新用户状态",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/user/{id}/unlock": {
      "post": {
        "operationId": "UserController_unlock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "解锁成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "解锁用户账号",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/user/{id}/reset-password": {
      "post": {
        "operationId": "UserController_adminResetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "mustChangePassword": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "重置成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "管理员重置用户密码并可选强制改密",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/user/{id}/roles": {
      "patch": {
        "description": "替换用户全部角色；会 bump 权限版本并强制下线目标用户。需要二次确认令牌。",
        "operationId": "UserController_assignRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "用户 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleIds"
                ],
                "properties": {
                  "roleIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "分配成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoSuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "覆盖式分配用户角色",
        "tags": [
          "用户管理"
        ]
      }
    },
...
    "/api/user/hard": {
      "delete": {
        "description": "从数据库永久移除用户记录，不可恢复；禁止删除当前登录用户自身。",
        "operationId": "UserController_hardRemoveMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUsersSwaggerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListItemArraySuccessSwaggerDto"
                }
              }
            }
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "批量物理删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Zen",
    "description": "Admin Zen API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "在 Authorization 头中传入 access token，格式：Bearer <token>"
      }
    },
    "schemas": {
      "CreateUserSwaggerDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "登录用户名，3–30 个字符",
            "minLength": 3,
            "maxLength": 30,
            "example": "zhangsan"
          },
          "email": {
            "type": "string",
            "description": "登录邮箱",
            "example": "zhangsan@example.com"
          },
          "password": {
            "type": "string",
            "description": "登录密码：至少 8 位，且包含大写、小写、数字与特殊字符",
            "minLength": 8,
            "example": "SecureP@ss1"
          },
          "nickname": {
            "type": "string",
            "description": "显示昵称，最长 50 个字符",
            "maxLength": 50,
            "example": "张三"
          },
          "phoneNumber": {
            "type": "string",
            "description": "手机号码，最长 20 个字符",
            "maxLength": 20,
            "example": "13800138000"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
...
      "CreateUserSuccessSwaggerDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务状态码（与 HTTP 状态码一致）",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "成功提示",
            "example": "Success"
          },
          "traceId": {
            "type": "string",
            "description": "请求追踪 ID（与响应头 x-trace-id 一致）",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间（ISO 8601）",
            "example": "2026-05-22T08:00:00.000Z"
          },
          "data": {
            "$ref": "#/components/schemas/UserListItemSwaggerDto"
          }
        },
        "required": [
          "code",
          "message",
          "traceId",
          "timestamp",
          "data"
        ]
      },
...
      "DeleteUsersSwaggerDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "要操作的用户 ID 列表，至少 1 个",
            "minItems": 1,
            "example": [
              "f47ac10b-58cc-4372-a567-0e02b2c3d479"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
    }
  }
}